+++ pyproject.toml
@@ -12,6 +12,7 @@ dependencies = [
     "langchain>=0.3.26",
     "langchain-community>=0.3.26",
     "langchain-ollama>=0.3.3",
+    "langchain-tavily>=0.2.10",
     "pydantic>=2.11.7",
     "scipy>=1.15.3",
     "sounddevice>=0.5.2",
